+++ tests/test_function/test_parse_scope.py
@@ -45,3 +45,16 @@ def test_parse_subscription() -> None:
 
     # then
     assert_that(result).is_equal_to(expected_result)
+
+
+def test_parse_tenant() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "tenant"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path=file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)